# Conflicts:
#	packages/library-base/src/commonMain/kotlin/io/realm/Configuration.kt
#	packages/library-base/src/commonMain/kotlin/io/realm/internal/Mediator.kt
#	packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmImpl.kt
#	packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmObjectHelper.kt
#	packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmObjectInternal.kt
#	packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmReference.kt
#	packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmUtils.kt
#	packages/library-base/src/commonMain/kotlin/io/realm/internal/schema/RealmSchemaImpl.kt
#	packages/library-base/src/jvmMain/kotlin/io/realm/internal/platform/SystemUtilsJvm.kt
#	packages/library-base/src/macosMain/kotlin/io/realm/internal/platform/SystemUtils.kt

+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmListInternal.kt
@@ -185,7 +185,7 @@ internal class ListOperator<E>(
         return with(metadata) {
             when (clazz) {
                 Byte::class -> (value as Long).toByte()
-                Char::class -> (value as Long).toChar()
+                Char::class -> (value as Long).toInt().toChar()
                 Short::class -> (value as Long).toShort()
                 Int::class -> (value as Long).toInt()
                 Long::class,
